fixed color_class

+++ opengl/source/color_class.cpp
@@ -2,7 +2,7 @@
 
 using vvis::visualization::VvisColor_3f;
 
-VvisColor_3f vvis::visualization::get_color_by_direction(int x, int y, int z) {
+VvisColor_3f vvis::visualization::get_color_by_direction(float x, float y, float z) {
 	// 
 	return VvisColor_3f(0, 0, 0);
 }
